refactor(dot11): extract frame subtype enums into subtype module
